backend/requirements/base.txt
Django==3.2.14
django-model-utils==4.2.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
PyJWT==1.6.4
django-cors-headers==3.13.0
django-filter==2.4.0
django-cities-light==3.9.2
raven==6.0.0
psycopg2==2.8.4
sentry-sdk==1.5.4
rest-condition==1.0.3
xlrd==1.2.0
Pillow==9.0.1
drf-renderer-xlsx==0.4.1
requests==2.22.0
django-rest-choices==0.0.1
reportlab==3.6.6
xhtml2pdf==0.2.5
openpyxl==3.0.10
